Assembly "hello world" applications.
To test our debugger we want  something "testable". ie: something that has a number
of instructions we can easily verify. When using CPP or a normal FPC program, they
include other units like System.pas etc. Thus the number of instructions in the relatively
simply program is still in the thousands. Using Assembly, we can reduce that to 7
instructions. So now we can easily verify our debugger by stepping through the whole
program.